flet_app/modules/minimax-remover/requirements.txt
decord==0.6
diffusers
pillow==9.2
einops==0.8.0
scipy==1.14.0
numpy==1.26.4
opencv-python==4.5.5.64
huggingface_hub==0.32.4
omegaconf
einops
accelerate==0.30.1
